+++ environment_tests/test-exports-cjs/src/entrypoints.js
@@ -18,7 +18,9 @@ const tools_render = require("langchain/tools/render");
 const tools_google_places = require("langchain/tools/google_places");
 const chains = require("langchain/chains");
 const chains_combine_documents_reduce = require("langchain/chains/combine_documents/reduce");
+const chains_history_aware_retriever = require("langchain/chains/history_aware_retriever");
 const chains_openai_functions = require("langchain/chains/openai_functions");
+const chains_retrieval = require("langchain/chains/retrieval");
 const embeddings_base = require("langchain/embeddings/base");
 const embeddings_cache_backed = require("langchain/embeddings/cache_backed");
 const embeddings_fake = require("langchain/embeddings/fake");
