+++ environment_tests/test-exports-esbuild/src/entrypoints.js
@@ -18,7 +18,9 @@ import * as tools_render from "langchain/tools/render";
 import * as tools_google_places from "langchain/tools/google_places";
 import * as chains from "langchain/chains";
 import * as chains_combine_documents_reduce from "langchain/chains/combine_documents/reduce";
+import * as chains_history_aware_retriever from "langchain/chains/history_aware_retriever";
 import * as chains_openai_functions from "langchain/chains/openai_functions";
+import * as chains_retrieval from "langchain/chains/retrieval";
 import * as embeddings_base from "langchain/embeddings/base";
 import * as embeddings_cache_backed from "langchain/embeddings/cache_backed";
 import * as embeddings_fake from "langchain/embeddings/fake";
