+++ environment_tests/test-exports-vercel/src/entrypoints.js
@@ -18,7 +18,9 @@ export * from "langchain/tools/render";
 export * from "langchain/tools/google_places";
 export * from "langchain/chains";
 export * from "langchain/chains/combine_documents/reduce";
+export * from "langchain/chains/history_aware_retriever";
 export * from "langchain/chains/openai_functions";
+export * from "langchain/chains/retrieval";
 export * from "langchain/embeddings/base";
 export * from "langchain/embeddings/cache_backed";
 export * from "langchain/embeddings/fake";
