+++ langchain/src/chains/tests/history_aware_retriever.test.ts
@@ -3,9 +3,9 @@ import { ChatPromptTemplate } from "@langchain/core/prompts";
 import { FakeRetriever } from "@langchain/core/utils/testing";
 import { Document } from "@langchain/core/documents";
 import { FakeListLLM } from "../../llms/fake.js";
-import { createChatHistoryRetriever } from "../conversational_retrieval_chain.js";
+import { createHistoryAwareRetriever } from "../history_aware_retriever.js";
 
-test("createChatHistoryRetriever", async () => {
+test("createHistoryAwareRetriever", async () => {
   const answer = "I know the answer!";
   const questionGenPrompt = ChatPromptTemplate.fromTemplate(
     `hi! {input} {chat_history}`
@@ -18,24 +18,14 @@ test("createChatHistoryRetriever", async () => {
   });
   const llm = new FakeListLLM({ responses: [answer] });
   const input = "What is the answer?";
-  const chain = createChatHistoryRetriever({
+  const chain = createHistoryAwareRetriever({
     llm,
     retriever,
     prompt: questionGenPrompt,
   });
   const output = await chain.invoke({ input, chat_history: [] });
-  expect(output).toEqual({
-    answer,
-    chat_history: [],
-    context: fakeRetrievedDocs,
-    input,
-  });
+  expect(output).toEqual(fakeRetrievedDocs);
 
   const output2 = await chain.invoke({ input, chat_history: "foo" });
-  expect(output2).toEqual({
-    answer,
-    chat_history: "foo",
-    context: fakeRetrievedDocs,
-    input,
-  });
+  expect(output2).toEqual(fakeRetrievedDocs);
 });
