+++ langchain/src/chains/tests/retrieval_qa.test.ts
@@ -2,7 +2,7 @@ import { test } from "@jest/globals";
 import { ChatPromptTemplate } from "@langchain/core/prompts";
 import { FakeRetriever } from "@langchain/core/utils/testing";
 import { Document } from "@langchain/core/documents";
-import { createRetrievalChain } from "../retrieval_qa.js";
+import { createRetrievalChain } from "../retrieval.js";
 import { FakeListLLM } from "../../llms/fake.js";
 
 test("createRetrievalChain", async () => {
