+++ langchain/src/load/import_map.ts
@@ -19,7 +19,9 @@ export * as tools__render from "../tools/render.js";
 export * as tools__google_places from "../tools/google_places.js";
 export * as chains from "../chains/index.js";
 export * as chains__combine_documents__reduce from "../chains/combine_documents/reduce.js";
+export * as chains__history_aware_retriever from "../chains/history_aware_retriever.js";
 export * as chains__openai_functions from "../chains/openai_functions/index.js";
+export * as chains__retrieval from "../chains/retrieval.js";
 export * as embeddings__base from "../embeddings/base.js";
 export * as embeddings__cache_backed from "../embeddings/cache_backed.js";
 export * as embeddings__fake from "../embeddings/fake.js";
